+++ src/wizard/Wizard.jsx
@@ -1,24 +1,28 @@
 import React from "react";
-import { ShallowComponent, Application } from "robe-react-commons";
+import {ShallowComponent, Application} from "robe-react-commons";
 import Col from "react-bootstrap/lib/Col";
 import Pager from "react-bootstrap/lib/Pager";
-import PageItem from "react-bootstrap/lib/PageItem";
 import Toast from "../toast/Toast";
+import Step from "./Step"
 import Button from "react-bootstrap/lib/Button";
 import FaIcon from "../faicon/FaIcon";
 import "./Wizard.css";
 
 export default class Wizard extends ShallowComponent {
     /**
-    * PropTypes of the component.
-    *
-    * @static
-    */
+     * PropTypes of the component.
+     *
+     * @static
+     */
     static propTypes = {
         /**
          * Current page index to render.
          */
-        currentStep: React.PropTypes.number,
+        currentKey: React.PropTypes.oneOfType([React.PropTypes.number, React.PropTypes.string]),
+        /**
+         * provides to wizard change own state.
+         */
+        changeState: React.PropTypes.bool,
         /**
          * Text for the next button.
          */
@@ -30,123 +34,120 @@ export default class Wizard extends ShallowComponent {
         /**
          * Text for the complete button.
          */
-        complete: React.PropTypes.string
+        complete: React.PropTypes.string,
+        /**
+         *
+         */
+        onCompleteClick: React.PropTypes.func,
+        /**
+         *
+         */
+        onChange: React.PropTypes.func
+
     };
 
     static defaultProps = {
-        currentStep: 0,
+        currentKey: undefined,
+        changeState: true,
         next: Application.i18n(Wizard, "wizard.Wizard", "next"),
         previous: Application.i18n(Wizard, "wizard.Wizard", "previous"),
         complete: Application.i18n(Wizard, "wizard.Wizard", "complete")
     };
-    __stepValidInfo = [];
-    __content = undefined;
+    __steps = [];
+    __refs = {};
 
-
-    constructor(props: Object) {
+    constructor(props:Object) {
         super(props);
-        this.state = {
-            currentStep: this.props.currentStep,
-            valid: true
-        };
+        this.componentWillReceiveProps(props);
     }
 
-    render(): Object {
+    componentWillReceiveProps(nextProps) {
+        this.state = {
+            currentKey: this.props.currentKey,
+            stateOfSteps: {},
+            valid: true
+        };
+        this.__initSteps(nextProps);
+    }
+
+    render():Object {
         return (
-            <div>
-                <Col className="wizard">
-                    <Col className="wizard-row setup-panel">
-                        {this.__renderSteps()}
-                    </Col>
-                </Col>
-                {this.__renderContent()}
+            <div id="wizard">
+                {this.__renderStepLayout()}
+                {this.__renderSteps()}
                 {this.__renderPager()}
             </div>
         );
     }
 
-    __renderSteps(): Array {
+    __renderStepLayout() {
         let steps = [];
-        for (let i = 0; i < this.props.steps.length; i++) {
-            let item = this.props.steps[i];
-            let styleClass = (this.state.currentStep === i) ? "btn-primary" : "btn-default";
-            let step = (
-                <Col key={i} className="wizard-step">
-                    <a
-                        onClick={() => { this.__onClickStepButton(i); } }
-                        className={`btn btn-circle ${styleClass}`}
-                        >
-                        {i + 1}
-                    </a>
-                    <p>{item.title}</p>
-                </Col>);
-            steps.push(step);
-        }
-        return steps;
+        let currentStep = this.__getCurrentStep();
+        this.__steps.map((child, idx)=> {
+            let previous = currentStep.index >= idx ? "previous-step" : "";
+            steps.push(
+                <div
+                    key={child.props.stepKey}
+                    className={`wizard-step ${previous}`}>
+                    <button
+                        id={(idx+1)}
+                        onClick={()=>this.__onChange(child.props)}
+                        className="wizard-step-button">
+                    </button>
+                    <br/>
+                    {child.props.title}
+                </div>)
+        });
+        return (
+            <div className="wizard">
+                <div className="wizard-row">
+                    {steps}
+                </div>
+            </div>);
     }
 
-    __onClickStepButton(index: number) {
-        if (this.state.currentStep === index) {
+    __renderSteps() {
+        let childArr = [];
+        this.__steps.map((child)=> {
+            let display = this.state.currentKey === child.props.stepKey ? "inherit" : "none";
+            childArr.push(
+                <div
+                    key={child.props.stepKey}
+                    style={{ display }}>
+                    {child}
+                </div>)
+        });
+        return childArr;
+    }
+
+    __renderPager():Object {
+        if (this.__steps.length <= 0) {
             return;
         }
-        this.__stepValidInfo[this.state.currentStep] = this.isValid();
-
-        let state = {};
-        if (this.state.currentStep < index) {
-            if (this.__areStepsValid(this.state.currentStep, index)) {
-                state.currentStep = index;
-                this.setState(state);
-            }
-        } else if (this.state.currentStep > index) {
-            state.currentStep = index;
-            this.setState(state);
-        }
-    }
-
-    __areStepsValid(start: number, end: number): boolean {
-        for (start; start < end; start++) {
-            if (!this.__stepValidInfo[start]) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    __renderContent(): Object {
-        this.__content = this.props.steps[this.state.currentStep].component;
-        return this.__content;
-    }
-
-    __handleNextButtonClick() {
-        this.__onClickStepButton(this.state.currentStep + 1);
-    }
-
-    __handlePreviousButtonClick() {
-        this.__onClickStepButton(this.state.currentStep - 1);
-    }
-
-    __renderPager(): Object {
+        let currentStep = this.__getCurrentStep();
         let nextButton = null;
-        if (this.state.currentStep === this.props.steps.length - 1) {
-            nextButton = (
-                <Col className="pull-right">
-                    <Button
-                        bsStyle="primary"
-                        onClick={this.props.onCompleteClick}
+        if (currentStep.index === this.__steps.length - 1) {
+            if (this.props.onCompleteClick) {
+                nextButton = (
+                    <Col className="pull-right">
+                        <Button
+                            bsStyle="primary"
+                            onClick={this.__onCompleteClick}
                         >
-                        <FaIcon code="fa-check-circle" />
-                        {this.props.complete}
-                    </Button>
-                </Col>
-            );
+                            <FaIcon code="fa-check-circle"/>
+                            {this.props.complete}
+                        </Button>
+                    </Col>
+                );
+            }
         } else {
             nextButton = (
                 <Pager.Item
                     next
                     disabled={!this.state.valid}
                     onClick={this.__handleNextButtonClick}
-                    >
-                    {this.props.next} <FaIcon code="fa-arrow-right" />
+                >
+                    {this.props.next} <FaIcon code="fa-arrow-right"/>
                 </Pager.Item>
             );
         }
@@ -154,29 +155,132 @@ export default class Wizard extends ShallowComponent {
             <Pager>
                 <Pager.Item
                     previous
-                    disabled={this.state.currentStep === 0}
-                    onClick={this.state.currentStep === 0 ? null : this.__handlePreviousButtonClick}
-                    >
-                    <FaIcon code="fa-arrow-left" />{this.props.previous}
+                    style={{display:currentStep.index==0?"none":"inherit"}}
+                    onClick={this.__handlePreviousButtonClick}
+                >
+                    <FaIcon code="fa-arrow-left"/>{this.props.previous}
                 </Pager.Item>
                 {nextButton}
             </Pager>
         );
     }
 
-    isValid(): boolean {
-        if (this.__content === undefined) {
-            return false;
+    __handleNextButtonClick() {
+        let currentStep = this.__getCurrentStep();
+        this.__onChange(this.__steps[currentStep.index + 1].props);
+    }
+
+    __handlePreviousButtonClick() {
+        let currentStep = this.__getCurrentStep();
+        this.__onChange(this.__steps[currentStep.index - 1].props);
+    }
+
+    __getCurrentStep() {
+        let currentStep = this.__steps[0] ? this.__steps[0].props : undefined;
+        this.__steps.map((child)=> {
+            if (child.props.stepKey === this.state.currentKey) {
+                currentStep = child.props;
+            }
+        });
+        return currentStep;
+    }
+
+    __initSteps(props) {
+        this.__steps = React.Children.map(props.children,
+            (child, idx) => {
+                if (child.type === Step) {
+                    return React.cloneElement(child, {
+                        title: child.props.title || "",
+                        stepKey: child.props.stepKey || ("step" + (idx + 1)),
+                        index: idx,
+                        ref: (step)=> {
+                            if (step) {
+                                this.__refs[step.props.stepKey] = step;
+                            }
+                        }
+                    });
+                }
+            });
+        this.__steps = this.__steps ? this.__steps : [];
+        if (this.state.currentKey === undefined) {
+            this.state.currentKey = this.__steps[0].props.stepKey;
         }
-        if (!this.__content._owner) { // eslint-disable-line no-underscore-dangle
-            return true;
+    }
+
+    __onCompleteClick() {
+        if (this.props.onCompleteClick) {
+            let currentStep = this.__getCurrentStep();
+            let currentStepRef = this.__refs[currentStep.stepKey];
+
+            let result = currentStepRef.isValid();
+            if (result.message) {
+                this.__showToast(result.type, result.message);
+            }
+            if (!result.status) {
+                return;
+            }
+
+            let stateOfSteps = this.state.stateOfSteps;
+            stateOfSteps[currentStep.stepKey] = currentStepRef.__getStateOfStep();
+            this.props.onCompleteClick(stateOfSteps)
+        }
+    }
+
+    __onChange(step) {
+        let currentStep = this.__getCurrentStep();
+        if (currentStep.stepKey === step.stepKey) {
+            return;
+        }
+        let isNext = step.index > currentStep.index;
+        let nextStep = isNext ? this.__steps[currentStep.index + 1].props : step;
+
+        if (isNext) {
+            let result = this.__refs[currentStep.stepKey].isValid();
+            if (result.message) {
+                this.__showToast(result.type, result.message);
+            }
+            if (!result.status) {
+                return;
+            }
         }
 
-        let result = this.__content._owner._instance.refs.step.isValid(); // eslint-disable-line no-underscore-dangle
-        if (!result) {
-            Toast.error(result.message);
-            return false;
+        if (this.props.changeState) {
+            let currentStepRef = this.__refs[currentStep.stepKey];
+            let nextStepRef = this.__refs[nextStep.stepKey];
+
+            let state = {currentKey: nextStep.stepKey};
+            let stateOfSteps = this.state.stateOfSteps;
+            stateOfSteps[currentStep.stepKey] = currentStepRef.__getStateOfStep();
+
+            this.setState(state, ()=> {
+                nextStepRef.__stateOfSteps(stateOfSteps)
+            });
         }
-        return true;
+
+        if (this.props.onChange) {
+            this.props.onChange({
+                target: nextStep
+            });
+        }
+    }
+
+    __showToast(type, message) {
+        switch (type) {
+            case "error":
+                Toast.error(message);
+                break;
+            case "info":
+                Toast.info(message);
+                break;
+            case "success":
+                Toast.success(message);
+                break;
+            case "warning":
+                Toast.warning(message);
+                break;
+            default:
+                Toast.error(message);
+                break;
+        }
     }
 }