+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.21 -- Developing...
 * Screen Keyboard component added.
+* Wizard component redeveloped.
 
 ## 1.1.20
 * Chart and Toast components test rewrited, Charts props updated